+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Add `generate_test_resources_from_file` script
   * Refactor custom `Matcher`s to be more readable
   * Create test & test data for web pipeline
+* Add warning to `BakeTableColumns`
 * Add new extension `ruby-lsp`
 * Create new version of `references` in form of footnotes
 * Add `BakeLinks` to nursing-external recipes